Adapting tests.

+++ data_adapter_oemof/adapters.py
@@ -77,12 +77,12 @@ class Adapter:
             # want to move this section including if statements together with decommissioning
             # section to a calculations as "default calculations
             if not isinstance(defaults["lifetime"], collections.abc.Iterable):
-                defaults["lifetime"] = np.floor(defaults["lifetime"])
+                defaults["lifetime"] = int(np.floor(defaults["lifetime"]))
             elif all(x == defaults["lifetime"][0] for x in defaults["lifetime"]):
-                defaults["lifetime"] = np.floor(defaults["lifetime"][0])
+                defaults["lifetime"] = int(np.floor(defaults["lifetime"][0]))
             else:
                 warnings.warn("Lifetime cannot change in Multi-period modeling")
-                defaults["lifetime"] = np.floor(defaults["lifetime"][0])
+                defaults["lifetime"] = int(np.floor(defaults["lifetime"][0]))
 
 
         return defaults
